requirements.txt
tensorflow-gpu~=2.10.0
numpy~=1.26.4
keras~=2.10.0
pandas~=2.2.2